[pkg-export-docker] Extract `RegistryType` to a non-CLI type.
This is a refactoring change which removes the CLI-only enum
`RegistryType` and makes it a proper business logic type in the root of
the `export_docker` module.

The underlying reason for this was twofold:

1) The default `Debug` and `Display` derivations would return the enum
variants as capitalized strings, and not lowercased. This causes the CLI
usage to look incorrect as we're looking for values like `azure` and
not `Azure`. Controlling the `FromStr` and `Display` implementations
yields a much better result.
2) This type was being imported by the rest of the module and used in
the other types and business logic which breaks the CLI/business logic
separation of concerns. If we need a type in non CLI-parsing code, it
shouldn't live in the CLI-parsing code.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>